Brings the plugin engine (PR #83) onto the branch that carries the Krait
sub-10 renderer work and the UI-emit overlay motion retention, so both land
in v3 together.

Resolutions:
- makefile: test-uitree sources are the union (overlay_retain from this
  branch; widget_anchor, widget_skin, frame_provide from the engine;
  frame_depth stays removed).
- uitree_emit.h: the overlay scratch fields stay; the declared-slot
  frame_order hints the engine removed are not re-added, and the overlay
  refresh no longer resets them (uitree_emit_overlay.u.h).
- uitree_emit_overlay.u.h: root validity reads widget_hidden in place of the
  removed replacement flags; the insert-at fixup runs over the two remaining
  overlay sources; emit_walk_node lost its role-group arguments.
- uitree_test_overlay_retain.c: the frame-overlay permutation becomes a
  canvas-overlay one (the frame overlay source is gone).

Gate: OPT=1 EMBED_SERVER=1 build in a private objdir; test-uitree,
test-cs2-highlight, test-minimenu-world, the plugin, gameframe, orbs, cs1,
transmit-pump and host-request manifest targets pass on the committed tree.

+++ src/ui/test/uitree_test_hover.c
@@ -563,7 +563,7 @@ test_hover_input(void)
         int32_t parent =
             UITree_TestPushXy(ht, -1, UIELEM_RS_LAYER, parent_id, 0, 0, 200, 200);
         int32_t old = UITree_CcCreate(ht, parent, parent_id, UIELEM_RS_RECT, 0);
-        uint32_t old_incarnation;
+        uint64_t old_incarnation;
         int saw_old_over = 0;
         int saw_new_over = 0;
         int saw_replacement_leave = 0;
